Added validators and refactoring for Follow and Identity services.

+++ src/Services/Identity/InstaConnect.Identity.Business/Commands/Account/ConfirmAccountEmail/ConfirmAccountEmailCommandHandler.cs
@@ -44,6 +44,11 @@ public class ConfirmAccountEmailCommandHandler : ICommandHandler<ConfirmAccountE
             throw new TokenNotFoundException();
         }
 
+        if (existingToken.UserId != request.UserId)
+        {
+            throw new AccountForbiddenException();
+        }
+
         _tokenRepository.Delete(existingToken);
 
         await _userRepository.ConfirmEmailAsync(existingUser.Id, cancellationToken);
